Practical 4 Invoke Workflow file
Create a simple interest calculator using a separate workflow and arguments

Create a simple interest calculator. In your ‘Main’ workflow, ask the user to enter the amount of the initial deposit and the period (multiple choice, e.g. 1 year, 3 years, 5 years). Store the input in two variables. Afterwards, create a third variable that will store the value of the final earning.

Create a new workflow and use it to calculate a simple interest using the user input and take into consideration the value of the rate/year. Afterwards, display on screen the cumulated interest at the end of the period and the final deposit balance.

Note: The formula to calculate the interest can be 'Principal amount x Rate per year x Period chosen / 100'. Use a ‘Switch’ activity to calculate the interest for all the period options (e.g: 1 year, 3 years, 5 years) and assign the result to an argument.