App/Web Recorder
This method of automating user interfaces is by using the Web/App Recorder feature. This is a tool that can help to save a lot of time when automating the business processes. This functionality enables to easily capture the actions on the screen and translate them into activities in Studio.