Practice 1 Input activities and methods
Input data into Notepad

Create a sequence that does the following Notepad automation:

Asks the user for the text to be typed.
Asks the user for the title under which the file should be saved.
Opens Notepad.
Types the user’s text.
Changes the font style to bold and its size to 16.
Navigates to Save as to save the file with the user’s title.
Closes Notepad.