Practice 1 Calculating Sum
Calculate the sum of the values from two columns in an Excel file in 3 ways. Create a workflow that adds the values on column A with values on column B and writes them in column C in different ways:

Keeps the Excel open and writes the results in real time, row by row so you can see the changes.
Keeps the Excel closed, set the column values in the memory DataTable and adds all the table to a new Excel files at once, in the end.
Calculates the sum by using Excel formulas in the original file.