Practice 5 Input validation using FlowDecision
Develop a workflow to ask for user text input according to a set of rules and incorporate the rules of input validation in the workflow.

More specifically, the string that the user inputs has to pass the following validation rules:

Minimum length: 8 characters
Starts with uppercase letter
Maximum 20 characters
The user is allowed 3 attempts to write the correct string (not taking into accounts the situations when the user typed nothing).

Note: Use a flowchart for this exercise.