Practice 6 Replacing Placeholders
Get input information and use it to fill in the placeholders in a string

You have the following template: "Hello Mr <first_name> <last_name>, we would like to invite you to our opening event next week on <day_of_week>. Please confirm by the end of current week.

Create a workflow to get input from the user and replace the provided input data in the phrase above.

Note: Create a string variable with the text above.