Practice_2_Set Data
Create a workflow that fills in the fields of a website with organized data that will be taken from an Excel file provided herein. It contains the following data: CompanyName, RoleInCompany, Address, Email, FullName, PhoneNumber.

Notes:
The first and last name has to be obtained from the full name column of the excel file.
Use only anchor-based selectors for this exercise.