Practice_1_Exercise_Debugging_and_Exception_Handling(project-Updated)
Debug your colleague's workflow (Updated using UiAutomationNext and data table)

Your colleague was tasked with creating a Proof of Concept that can demonstrate UiPath's adaptability to changing interfaces. The robot needs to extract the first name, last name, phone number, and company name from www.fakenamegenerator.com and insert them in the correct fields on rpachallenge.com.

The robot is not behaving correctly and your colleague needs your help to make it work. Can you download the project and debug this workflow using the methods discussed in the first chapter?