RetryScope - Demo
The Retry Scope activity retries the contained activities as long as the condition is not met or an error is thrown. The Retry Scope activity is used for catching and handling an error, which is why it’s similar to TryCatch. The difference is that this activity simply retries the execution instead of providing a more complex handling mechanism. Filling in the details and downloading an ACME report.