Practice 1 - Automation of Spreadsheet Calculation - Updated
Handle Exceptions in your client's Cash Flow

Your task is to automate the calculation of the difference between two fields in an Excel file (Cash In and Cash Out) and input it in the third column. Unfortunately, the file has accumulated a lot of errors over time. A simple workflow will throw System.Exception because the robot will try to perform math calculations with Int32 variables that contain letters.

Create a workflow that takes as input the excel file attached below and adds the difference between Cash In value and Cash Out value for each row on the Difference column. If the calculation of the difference between the two is successful, set the Status to Success, otherwise set it to either Cash In wrong or Cash Out wrong.

Use Try Catch blocks to catch exceptions when trying to convert Cash In and Cash Out values.