Practice1_Organize Emails
Project that reads the Inbox and sorts the emails by moving them to various folders, based on “rules” defined in an Excel similar to the one below.

If the sender address contains the String defined in the Sender column, move the mail to the the email folder defined in the Folder column.