Practice 2 Send Email by outlook with attachment
You are receiving emails on a daily basis with details of newly hired persons. Each new employee needs an unique ID number to be generated for them. Since this is a repetitive task that is taking precious hours from you, why not automate it?

To do so, start a new process in Studio, get the mail messages from your 'Inbox' folder that have 'IDs Request' as subject, store the sender's email address, then save the attachments (Excel file). Next, you must:

1. Read the Excel file.

2. Generate an unique ID from the first three letters of the first name, first three letters of the last name, and last three digits of the phone number.

3. Add all the new IDs to the ID column previously created.

4. Send an Outlook email message based on a template and with the attached updated Excel file back to the initial sender. The template must contain the original sender's name, the task completion time and your name.