Practice 1 - Connect an Unattended Robot to Orchestrator
You're probably set with your attended environment. Studio is installed and licensed and you have an attended robot connected to Orchestrator via interactive sign-in, and controlled through UiPath Assistant.

It's time to take your environment further by provisioning an unattended robot and connecting in to the same Orchestrator. You need a separate machine to deploy it and an Orchestrator with an unattended license (which could be a community Automation Cloud).