Practice 2 - Orchestrator Resources
It's time to take the project we developed earlier a bit further. As a general principle, we should aim to keep static values out of our projects. The simple reason is that it's easier to update a value stored outside a workflow than ask the developer to update a project.

The request is straightforward: update the project below (the one we developed in the Working with Orchestrator resources in Studio lesson) to retrieve the ACME URL from Orchestrator.