Practice 3 - Queues
You are tasked with automating the process of inserting transactions in the teller application "UiDemo". The transaction-related input data is stored in an Excel file. Due to the very large number of transactions, multiple robots are required to work at the same time. According to the Solution Architect (SA) and the Business Analyst (BA), the process will need to be allocated to as many as 5 robots.